Add files via upload
This is a spreadsheet to track ongoing fire related incidents throughout the year. Information can be viewed on a yearly, quarterly or monthly basis with charts for year over year and month over month changes. All previous year and current year data is entered on a single worksheet and all related worksheets update accordingly.